refactor: 소소한 네이밍 수정
