+++ REQUIREMENTS.md
@@ -1,16 +1,21 @@
 ## 설문지 페이지 `/form/:id`
 
-- [ ] 설문지 데이터를 받아 동적으로 form들을 그린다.
-- [ ] form input들을 각 각 조작할 수 있다.
-- [ ] required 항목은 표시해주고 유효성 검사한다. (입력 안했을 시, 제출 불가)
-- [ ] 제출 버튼 클릭시, 설문지 응답 결과 노출한다.
-- [ ] 이전, 다음 버튼으로 섹션을 오갈 수 있다.
-  - [ ] 마지막 섹션 페이지에서는 "다음" 대신 "제출" 버튼 노출
-  - [ ] 첫 번째 섹션 페이지에서는 "이전" 버튼 비활성화 (혹은 미노출)
+- [x] 설문지 데이터를 받아 동적으로 form들을 그린다.
+- [x] form input들을 각 각 조작할 수 있다.
+- [x] required 항목은 표시해주고 유효성 검사한다. (입력 안했을 시, 제출 불가)
+- [x] 제출 버튼 클릭시, 설문지 응답 결과 노출한다.
+  - 제출 결과 페이지로 이동
+- [x] 이전, 다음 버튼으로 섹션을 오갈 수 있다.
+  - [x] 마지막 섹션 페이지에서는 "다음" 대신 "제출" 버튼 노출
+  - [x] 첫 번째 섹션 페이지에서는 "이전" 버튼 비활성화 (혹은 미노출)
+
+## 설문지 결과 페이지 `/form/:id/submit`
+
+- [x] formID에 맞는 설문지 결과를 불러온다.
 
 ## 추가 기능
 
-- [ ] 양식 지우기 기능
-- [ ] 불러오기 기능
+- [x] 양식 지우기 기능
+- [x] 불러오기 기능
   - 로컬 스토리지 / 세션 스토리지를 활용해 데이터 저장
   - 페이지 새로고침 시 저장된 데이터 불러오기
